+++ packages/core-cairo/src/utils/strings.ts
